pyproject.toml
[tool.poetry]
name = "ibis_async_app"
version = "0.1.0"
description = ""
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
license = "MIT"
# packages = [
#     { include = "./src/" },
# ]
include = [
  "*.cfg",
  "*.ini",
  "*.js",
  "*.json",
  "*.lock",
  "*.md",
  "*.py",
  "*.sh",
  "*.rst",
  "*.txt",
  "*.toml",
  "*.yml",
  "*.yaml",
  ".github",
  "conda",
  "docker",
  "docs",
  ".dockerignore",
  ".env.tpl",
  ".gitignore",
  "Makefile",
]
exclude = [
  ".git/*",
  ".env*",
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
ibis-framework = "^3.0.2"
fastapi = "^0.77.1"
uvicorn = "^0.17.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
